simpler routing tier
the updating history based on actions was cool and all, but prevents me
from being able to construct regular old <a href="..."> links.

this switches to elm-history and elm-router with an explicit Routes
module which encapsulates the parsing and construction of routes.

switched StartApp to a version with evancz/start-app#37 merged so that
loading the page goes to the route specified in the URL.

this also fixes the awkward coupling between Index and Norad; previously
Norad intercepted the Index.GoToPipeline action to know when to switch
pages. now it's all done through the Routes package and simply linking
to the route, which feeds back into Norad via history updates.