route '' to Index
this mainly affects going back from #!/... to the original page.
previously it would not be able to find the route.

this would have been caught sooner if an actual 404 page was
implemented, but right nwo 404s are treated as noops, and the default
page is the index, so nothing bad happened.